chore(explore) Addressing more explore notion points.
